moved 'GroomerDisplay.js' from 'components/pages/ProfileDisplay' to 'components/groomers'; updated GroomerDisplay import in SearchForm.js to reflect new file structure; deleted 'ProfileDisplay' directory because it is now empty
